% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_add_ranges}
\alias{Block_add_ranges}
\title{Store a ranges object in the Block store.}
\usage{
Block_add_ranges(Block = NULL, ranges = NULL, rangekey = NULL,
    remove.existing = TRUE)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{ranges}{\strong{Required}.
An object of class ranges specifying the ranges to store in the Block.}

\item{rangekey}{\strong{Required}.
The name to use for the ranges within the Block store.}

\item{remove.existing}{\strong{Optional}. TRUE
Will remove an existing Ranges by the same name and introduce the new one.}
}
\value{
Returns TRUE if completed successfully.
}
\description{
\code{Block_add_ranges} loads a GRanges object into the Block store.
}
\details{
With this function it is possible to associate other ranges objects with the
Block store. If metadata columns are present, the are also loaded into the
Block store. Although not explicitly asked for, the metadata columns should
not be of type list as this may create complications down the line. We
ask for ranges objects, so if the same ranges object is later retrieved
two additional columns will be present. These are the strand and width
columns that are obtained when a ranges is converted into a data.frame.
Users can ignore these columns.
}
\examples{
Bintable.path <- system.file("extdata",
"Bintable_40kb.txt", package = "HiCBlocks")
Chromosomes <- "chr19"
Path_to_cached_file <- CreateBlock(ChromNames = Chromosomes,
BinTable = Bintable.path, bin.delim = " ",
Output.Filename = "test.hdf", exec = "cat",
remove.existing = TRUE)

Block.file <- Block_path_to_file("test.hdf")
Chrom <- c("chrS","chrS","chrS","chrS","chrS")
Start <- c(10000,20000,40000,50000,60000)
End <- c(10001,20001,40001,50001,60001)
Test_ranges <- Block_make_ranges(Chrom = Chrom, Start = Start, End = End)
Block_add_ranges(Block = Block.file, ranges = Test_ranges,
rangekey = "test_ranges")

\dontrun{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Chrom <- c("chrS","chrS","chrS","chrS","chrS")
Start <- c(10000,20000,40000,50000,60000)
End <- c(10001,20001,40001,50001,60001)
Test_ranges <- Block_make_ranges(Chrom = Chrom, Start = Start, End = End)
Block_add_ranges(Block = Block.file, ranges = Test_ranges,
rangekey = "test_ranges", remove.existing = TRUE)

}

}
