% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_get_bintable}
\alias{Block_get_bintable}
\title{Returns the binning table associated to the Hi-C experiment.}
\usage{
Block_get_bintable(Block = NULL, chr = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr}{\strong{Optional}.
A chr string specifying the chromosome to select from the ranges.}
}
\value{
Returns a GRanges object containing the binning
table associated to the Block store.
}
\description{
\code{Block_get_bintable} makes a call to \code{\link{Block_get_ranges}} to
retrieve the binning table of the associated Block store. This is equivalent
to passing the argument rangekey = "bintable" in
\code{\link{Block_get_ranges}}
}
\examples{

Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_get_bintable(Block = Block.file)

}
\seealso{
Block_get_ranges
}
