% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_get_chrominfo}
\alias{Block_get_chrominfo}
\title{Get the chrominfo for the Hi-C experiment.}
\usage{
Block_get_chrominfo(Block = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}
}
\value{
A three column data.frame containing chromosomes, nrows and length.

chromosomes corresponds to all chromosomes in the provided bintable.

nrows corresponds to the number of entries in the bintable or dimension
for that chromosome in a Hi-C matrix.

Length is the total bp length of the same
chromosome (max value for that chromosome in the bintable).
}
\description{
\code{Block_get_chrominfo} fetches the associated chrominfo table for the
Block it is associated to.
}
\examples{
Block.file = system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_get_chrominfo(Block = Block.file)
}
