% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_get_matrix_mcols}
\alias{Block_get_matrix_mcols}
\title{Get the matrix metadata columns in the Block store.}
\usage{
Block_get_matrix_mcols(Block = NULL, chr1 = NULL, chr2 = NULL,
    what = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{what}{\strong{Required}
A character vector of length 1 specifying the matrix metric to retrieve}
}
\value{
Returns a 1xN dimensional vector containing the specified matrix
metric
}
\description{
\code{Block_get_matrix_mcols} will get the specified matrix metadata column.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_get_matrix_mcols(Block = Block.file, chr1 = "chr19", chr2 = "chr19",
what = "bin.coverage")
}
