% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_get_matrix_within_coords}
\alias{Block_get_matrix_within_coords}
\title{Return a matrix subset between two regions.}
\usage{
Block_get_matrix_within_coords(Block = NULL, x.coords = NULL,
    y.coords = NULL, force = FALSE, FUN = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{x.coords}{\strong{Required}.
A string specifying the region to subset on the rows. It takes the form
chr:start:end. An overlap operation with the associated bintable will be
done to identify the bins to subset on the row}

\item{y.coords}{\strong{Required}.
A string specifying the region to subset on the rows. It takes the form
chr:start:end. An overlap operation with the associated bintable will be
done to identify the bins to subset on the column}

\item{force}{\strong{Optional}. Default FALSE
If true, will force the retrieval operation when matrix contains loaded
data until a certain distance.}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before
the matrix is returned.}
}
\value{
Returns a matrix of dimension x.coords binned length by y.coords
binned length. This may differ based on FUN.
}
\description{
\code{Block_get_matrix_within_coords} will fetch a matrix subset after
creating an overlap operation between both regions and the bintable
associated to the Block store.
This function calls \code{\link{Block_get_matrix}}.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_get_matrix_within_coords(Block = Block.file,
x.coords = "chr19:40000:2000000",
y.coords = "chr19:40000:2000000")

Block_get_matrix_within_coords(Block = Block.file,
x.coords = "chr19:40000:2000000",
y.coords = "chr19:40000:2000000",
FUN = mean)

Block_get_matrix_within_coords(Block = Block.file,
x.coords = "chr19:40000:2000000",
y.coords = "chr19:40000:2000000",
FUN = median)

}
\seealso{
\code{\link{Block_get_matrix}} to get matrix by using matrix
coordinates, \code{\link{Block_get_values_by_distance}} to get values
separated at a certain distance, \code{\link{Block_fetch_row_vector}}
to get values in a certain row/col and subset them,
\code{\link{Block_get_vector_values}} to get values using matrix
coordinates.
}
