% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_is_tracked}
\alias{Block_is_tracked}
\title{Check if a Block is being tracked.}
\usage{
Block_is_tracked(Block = NULL)
}
\arguments{
\item{Block}{\strong{required}
Path to the Block file to check.}
}
\value{
Returns TRUE or FALSE indicating if the file path is being tracked
or not.
}
\description{
\code{Block_is_tracked} checks if the provided Block file is being tracked or not.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks") 
Block_is_tracked(Block = Block.file)
}
