% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_list_matrices}
\alias{Block_list_matrices}
\title{List the matrix pairs present in the Block store.}
\usage{
Block_list_matrices(Block = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}
}
\value{
Returns a data.frame object with columns chr1, chr2 corresponding
to chromosome pairs, and the associated attributes. filename corresponds to
the name of the file that was loaded for the pair. min and max specify the
minimum and maximum values in the matrix, done is a logical value
specifying if a matrix has been loaded and sparsity specifies if a matrix
is defined as a sparse matrix.
}
\description{
\code{Block_list_matrices} will list all chromosomal pair matrices from the Block
store, with their associated filename, value range, done status and sparse
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_list_matrices(Block = Block.file)

}
