% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_list_mcool_normalisations}
\alias{Block_list_mcool_normalisations}
\title{Get all available normalisations in an mcool file.}
\usage{
Block_list_mcool_normalisations(names.only = FALSE)
}
\arguments{
\item{names.only}{\strong{Optional}. Default FALSE
A parameter specifying whether to list only the human readable names without
their respective column names in the mcool file.}
}
\value{
A named vector listing all possible normalisation factors.
}
\description{
\code{Block_list_mcool_normalisations} lists the names available for
accessing the various normalisation factors in an mcool file. Please note,
this only lists the mapping of the columns to their respective names.
It does not check for the availability of that particular column in
the mcool file
}
\examples{
Block_list_mcool_normalisations()

}
