% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_list_rangekeys}
\alias{Block_list_rangekeys}
\title{List the ranges tables stored within the Block.}
\usage{
Block_list_rangekeys(Block = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}
}
\value{
A one dimensional character vector of length x specifying the names
of all ranges currently present in the file.
}
\description{
\code{Block_list_rangekeys} lists the names of all ranges associated to a Block.
}
\examples{

Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_list_rangekeys(Block = Block.file)

}
