% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_list_ranges_mcols}
\alias{Block_list_ranges_mcols}
\title{Find out what metadata columns are associated to a ranges with a certain
name}
\usage{
Block_list_ranges_mcols(Block = NULL, rangekey = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{rangekey}{\strong{Optional}.
A string specifying the name of the ranges. If not present, the metadata
columns of all ranges will be listed.}
}
\value{
if no metadata columns are present, NA. If metadata columns are
present, a data.frame object containing the name of the ranges and the
associated metadata column name.
}
\description{
\code{Block_list_ranges_mcols} will list the metadata columns of the specified
ranges if it is present in the Block store.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_list_ranges_mcols(Block = Block.file, rangekey = "test_ranges")

}
