% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_list_tracked_blocks}
\alias{Block_list_tracked_blocks}
\title{List all Blocks that have been created with .}
\usage{
Block_list_tracked_blocks(detailed = FALSE, preserve.col.names = FALSE)
}
\arguments{
\item{detailed}{\strong{Optional}
If FALSE, produces a data.frame of file paths pointing to the
Block objects. If TRUE, produces a tibble containing additional information
such as the bioc cache tracking id, and the creation, last accession and
modification times.}

\item{preserve.col.names}{\strong{Optional}
If TRUE, will preserve the original colnames from the bioc cache tibble
object. If FALSE, will attempt to humanize the colnames to improve
readability.}
}
\value{
Returns a data.frame or tibble containing the path to tracked
Block objects. If tibble will contain additional information related to
the tracking. For details see parameter detailed.
}
\description{
\code{Block_list_tracked_blocks} will list all Block objects that are being
tracked.
}
\examples{
Block_list_tracked_blocks()
}
