% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_matrix_exists}
\alias{Block_matrix_exists}
\title{Check if a chromosome pair exists.}
\usage{
Block_matrix_exists(Block = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns a logical vector of length 1, specifying if the matrix
exists or not.
}
\description{
Matrices are created when the bintable is loaded and the chromosome names
are provided. If a user is in doubt regarding whether a matrix is present or
not it is useful to check this function. If the Bintable did not contain a
particular chromosome, any matrices for that chromosome would not be present
in the file
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_matrix_exists(Block = Block.file, chr1 = "chr19", chr2 = "chr19") 

Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_matrix_exists(Block = Block.file, chr1 = "chr19", chr2 = "chr20")

}
