% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_matrix_filename}
\alias{Block_matrix_filename}
\title{Return the filename of the loaded matrix}
\usage{
Block_matrix_filename(Block = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns a character vector of length 1 specifying the filename of
the currently loaded matrix.
}
\description{
Return the filename of the loaded matrix
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_matrix_filename(Block = Block.file, chr1 = "chr19", chr2 = "chr19")

}
