% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_matrix_issparse}
\alias{Block_matrix_issparse}
\title{Check if a matrix for a chromosome pair is sparse.}
\usage{
Block_matrix_issparse(Block = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns a logical vector of length 1, specifying if a matrix was
loaded as a sparse matrix.
}
\description{
Check if a matrix for a chromosome pair is sparse.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_matrix_issparse(Block = Block.file, chr1 = "chr19", chr2 = "chr19")

}
