% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_matrix_maxdist}
\alias{Block_matrix_maxdist}
\title{Get the maximum loaded distance from the diagonal of any matrix.}
\usage{
Block_matrix_maxdist(Block = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns an integer vector of length 1, specifying the maximum
distance loaded for that matrix
}
\description{
If values beyond a certain distance were not loaded in the matrix, this
distance parameter is useful. This package by default will check this param
to make sure that it is not returning non-existent data.

\code{Block_matrix_maxdist} will return this parameter.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_matrix_maxdist(Block = Block.file, chr1 = "chr19", chr2 = "chr19") 

}
