% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_path_to_file}
\alias{Block_path_to_file}
\title{Check if a Block is being tracked.}
\usage{
Block_path_to_file(Block = NULL)
}
\arguments{
\item{Block}{\strong{required}
Path to the Block file to locate.}
}
\value{
Returns the corresponding path to the file that is being tracked.
}
\description{
\code{Block_path_to_file} will query the tracking cache and return the complete
path to the file that is being tracked.
}
\examples{
Block.file <- "test.hdf"
Block_path_to_file(Block = Block.file)
}
