% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_rangekey_exists}
\alias{Block_rangekey_exists}
\title{Check to see if the Block contains a ranges with a certain name.}
\usage{
Block_rangekey_exists(Block = NULL, rangekey = NULL)
}
\arguments{
\item{Block}{\strong{Required}.
A string specifying the path to the Block store created with CreateBlock.}

\item{rangekey}{\strong{Required}.
A string specifying the name of the ranges to check for.}
}
\value{
A logical vector of length 1 with either TRUE or FALSE values.
}
\description{
\code{Block_rangekey_exists} checks for the presence of a particular ranges with
a certain name.
}
\examples{

Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_rangekey_exists(Block = Block.file, rangekey = "Bintable")

}
