% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_track_blocks}
\alias{Block_track_blocks}
\title{Check if a Block is being tracked.}
\usage{
Block_track_blocks(Block = NULL)
}
\arguments{
\item{Block}{\strong{required}
Path to the Block file to track.}
}
\value{
Returns a named vector containing the path to the file along with
the bioc Cache id as its name.
}
\description{
\code{Block_track_blocks} will start tracking the provided Block file.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
if(Block_is_tracked(Block.file)) {
    Block_untrack_block(Block = Block.file)
    Block_track_blocks(Block = Block.file)
}else{
    Block_track_blocks(Block = Block.file)
}
}
