% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Block_functions.R
\name{Block_untrack_block}
\alias{Block_untrack_block}
\title{Check if a Block is being tracked.}
\usage{
Block_untrack_block(Block = NULL)
}
\arguments{
\item{Block}{\strong{required}
Path to the Block file to untrack.}
}
\value{
Returns any value provided by bfcremove.
}
\description{
\code{Block_untrack_block} will stop tracking the provided Block file,
if it exists.
}
\examples{
Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
if(Block_is_tracked(Block.file)){
    Block_untrack_block(Block = Block.file)
}else{
    Block_track_blocks(Block = Block.file)
    Block_untrack_block(Block = Block.file)
}
}
