% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_art_functions.R
\name{Block_vizart_plot_heatmap}
\alias{Block_vizart_plot_heatmap}
\title{Create the entire HDF5 structure and load the bintable}
\usage{
Block_vizart_plot_heatmap(File = NULL, Blocks = NULL,
    x.coords = NULL, y.coords = NULL, FUN = NULL, value.cap = NULL,
    distance = NULL, rotate = FALSE, x.axis = TRUE,
    x.axis.title = NULL, y.axis = TRUE, y.axis.title = NULL,
    title = NULL, legend.title = NULL, return.object = FALSE,
    x.axis.num.breaks = 5, y.axis.num.breaks = 5, palette = NULL,
    col.direction = 1, extrapolate.on = NULL, x.axis.text.size = 10,
    y.axis.text.size = 10, text.size = 10, legend.title.text.size = 8,
    legend.text.size = 8, title.size = 10, tad.ranges = NULL,
    group.col = NULL, tad.colour.col = NULL, colours = NULL,
    colours.names = NULL, cut.corners = FALSE, highlight.points = NULL,
    width = 10, height = 6, line.width = 0.5, units = "cm",
    legend.key.width = unit(3, "cm"), legend.key.height = unit(0.5,
    "cm"))
}
\arguments{
\item{File}{\strong{Required}
A character vector containing the output filename to write.}

\item{Blocks}{\strong{Required}
A character vector of length 1 (in case of one sample heatmaps) or 2 (in case
of two sample heatmaps) specifying the names of the Block stores from where
to fetch the data.}

\item{x.coords}{\strong{Required}
A character vector of length 1 specifying the coordinates from where to fetch
the data.}

\item{y.coords}{\strong{Required}
A character vector of length 1 specifying the coordinates from where to fetch
the data.}

\item{FUN}{\strong{Optional}. Default NULL
If any sort of transformations should be applied to the data before plotting.
Such as, log10 or log2 transformations.}

\item{value.cap}{\strong{Optional}. Default NULL
If present, values beyond a certain quantile will be capped to that quantile.
In Hi-C this helps to emphasize structural information. Please note, if this
parameter is present the greatest value will have a greater than sign append-
-ed to them.}

\item{distance}{\strong{Optional}. Default NULL
If present, values beyond this distance will be filtered out. Please note,
that if a Block store matrix was loaded until a certain distance, this
parameter will result in an error if it is greater than the loaded distance.}

\item{rotate}{\strong{Optional}. Default FALSE
If TRUE, will rotate the heatmap by 90 degrees.}

\item{x.axis}{\strong{Optional}. Default TRUE
If FALSE, the x-axis will be removed (ticks, x-axis labels and title).}

\item{x.axis.title}{\strong{Optional}. Default NULL
If present, will be the \emph{x-axis} title. Else defaults to the provided
x.coords}

\item{y.axis}{\strong{Optional}. Default TRUE
If FALSE, the y-axis will be removed (ticks, y-axis labels and title).}

\item{y.axis.title}{\strong{Optional}. Default NULL
If present, will be the \emph{y-axis} title. Else defaults to the provided
y.coords}

\item{title}{\strong{Optional}. Default NULL
If present, will be the \emph{plot} title. Else defaults to the provided
x.coords vs y.coords}

\item{legend.title}{\strong{Optional}. Default NULL
If present will be the title of the legend. Else defaults to "Signal".}

\item{return.object}{\strong{Optional}. Default FALSE
If present the ggplot object will be returned}

\item{x.axis.num.breaks}{\strong{Optional}. Default 5
Number of ticks on the x axis}

\item{y.axis.num.breaks}{\strong{Optional}. Default 5
Number of ticks on the y axis}

\item{palette}{\strong{Required}. Default NULL
One of the RColorbrewer or viridis colour palettes}

\item{col.direction}{\strong{Optional}. Default 1
If -1, the colour scale will be reversed.}

\item{extrapolate.on}{\strong{Optional}. Default NULL
If present, colours from the palette will be extrapolated between lightest
and darkest to create the gradient. This value cannot be more than 100.}

\item{x.axis.text.size}{\strong{Optional}. Default 10
x-axis text size}

\item{y.axis.text.size}{\strong{Optional}. Default 10
y-axis text size}

\item{text.size}{\strong{Optional}. Default 10
text size of text elements in the plot.}

\item{legend.title.text.size}{\strong{Optional}. Default 8
text size of the legend title}

\item{legend.text.size}{\strong{Optional}. Default 8
text size of the legend text}

\item{title.size}{\strong{Optional}. Default 10
text size of the title}

\item{tad.ranges}{\strong{Optional}. Default NULL
A GenomicRanges object specifying the start and end coordinates of TADs to be
plotted on the heatmap.}

\item{group.col}{\strong{Optional}. Default NULL
Name of the column which will be used to categorize TADs as belonging to
either the first or the second Block stores. This must be a numeric value
ranging from 1 to 2. If NULL, TADs will be plotted on both Hi-C maps.}

\item{tad.colour.col}{\strong{Optional}. Default NULL
tad.colour.col takes as value the column name in the tad.ranges object
corresponding to the column which should be used to define different TAD
categories.}

\item{colours}{\strong{Optional}. Default NULL
If tad.ranges is present, colours expects a hexcode value of length 1. But,
if tad.colour.col is specified, it expects colours of the same length as
unique tad.ranges$tad.colour.col.}

\item{colours.names}{\strong{Optional}. Default NULL
If present, will be assigned to colours. Else, will inherit unique
tad.colour.col. If tad.colour.col is also absent, will revert to a placehold
column name.}

\item{cut.corners}{\strong{Optional}. Default FALSE
if cut.corners is TRUE, TAD borders will not be truncated, and they will
span until the end of visible heatmap.}

\item{highlight.points}{\strong{Optional}. Not yet implemented.}

\item{width}{\strong{Optional}. Default 10cm
Width of the output file units.}

\item{height}{\strong{Optional}. Default 6cm
Height of the output file in units.}

\item{line.width}{\strong{Optional}. Default 0.5
When plotting TADs set the width of the plotted lines}

\item{units}{\strong{Optional}. Default cm
Defines the units of the output file width and height.}

\item{legend.key.width}{\strong{Optional}. Default unit(3,"cm")
Defines the legend key width.}

\item{legend.key.height}{\strong{Optional}. Default unit(0.5,"cm")
Defines the legend key height.}
}
\value{
If return.object is set to TRUE, the constructed ggplot2
object will be returned. Else TRUE.
}
\description{
\code{Block_vizart_plot_heatmap} creates various heatmaps and plots TADs.
}
\details{
This function provides the capability to plot various types of heatmaps from
Hi-C data.
\itemize{
\item One sample heatmap.
\item Two sample heatmap (One sample on upper and other on lower).
\item All of the above with 90 degree rotation.
\item All of the above but with signal capped at a certain value.
\item All of the above but filtered by distance.
\item All of the above with TADs/TAD borders plotted on top.
}
}
\examples{
FailSafe_log10 <- function(x){
    x[is.na(x) | is.nan(x) | is.infinite(x)] <- 0
    return(log10(x+1))
}

Block.file <- system.file("extdata", "test.hdf", package = "HiCBlocks")
Block_vizart_plot_heatmap(File = "./chr19-5000000-10000000.pdf", 
Blocks = Block.file, x.coords = "chr19:5000000:10000000", palette = "Reds",
y.coords = "chr19:5000000:10000000", FUN = FailSafe_log10, 
value.cap = 0.99, width = 10, height = 11, legend.key.width = unit(3,"mm"),
legend.key.height = unit(0.3,"cm"))

}
