% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_bintable}
\alias{Brick_get_bintable}
\title{Returns the binning table associated to the Hi-C experiment.}
\usage{
Brick_get_bintable(Brick, chr = NA, resolution = NA)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with
Create_many_Brick.}

\item{chr}{\strong{Optional}.
A chr string specifying the chromosome to select from the ranges.}

\item{resolution}{\strong{Optional}. Default NA
When an object of class BrickContainer is provided, resolution defines the
resolution on which the function is executed}
}
\value{
Returns a GRanges object containing the binning
table associated to the Brick store.
}
\description{
\code{Brick_get_bintable} makes a call to \code{\link{Brick_get_ranges}} to
retrieve the binning table of the associated Brick store. This is equivalent
to passing the argument rangekey = "bintable" in
\code{\link{Brick_get_ranges}}
}
\examples{

Container_file <- system.file(file.path("extdata"), "HiCBricks")
My_BrickContainer <- load_BrickContainer(project_dir = Container_file)
Brick_get_bintable(Brick = My_BrickContainer, resolution = 100000)

}
\seealso{
Brick_get_ranges
}
