% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_is_tracked}
\alias{Brick_is_tracked}
\title{Check if a Brick is being tracked.}
\usage{
Brick_is_tracked(Brick)
}
\arguments{
\item{Brick}{\strong{required}
Path to the Brick file to check.}
}
\value{
Returns TRUE or FALSE indicating if the file path is being tracked
or not.
}
\description{
\code{Brick_is_tracked} checks if the provided Brick file is being
tracked or not.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_is_tracked(Brick = Brick.file)
}
