% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_list_matrices}
\alias{Brick_list_matrices}
\title{List the matrix pairs present in the Brick store.}
\usage{
Brick_list_matrices(Brick = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}
}
\value{
Returns a data.frame object with columns chr1, chr2 corresponding
to chromosome pairs, and the associated attributes. filename corresponds to
the name of the file that was loaded for the pair. min and max specify the
minimum and maximum values in the matrix, done is a logical value
specifying if a matrix has been loaded and sparsity specifies if a matrix
is defined as a sparse matrix.
}
\description{
\code{Brick_list_matrices} will list all chromosomal pair matrices from the
Brick store, with their associated filename, value range, done status
and sparse
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_list_matrices(Brick = Brick.file)

}
