% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_matrix_dimensions}
\alias{Brick_matrix_dimensions}
\title{Return the dimensions of a matrix}
\usage{
Brick_matrix_dimensions(Brick, chr1, chr2, resolution = NA)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with
Create_many_Brick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{resolution}{\strong{Optional}. Default NA
When an object of class BrickContainer is provided, resolution defines the
resolution on which the function is executed}
}
\value{
Returns the dimensions of a Hi-C matrix for any given
chromosome pair.
}
\description{
Return the dimensions of a matrix
}
\examples{
Container_file <- system.file(file.path("extdata",
"HiCBricks_builder_config.json"), package = "HiCBricks")
Brick_test <- load_BrickContainer(Container_file)
Brick_matrix_dimensions(Brick = Brick_test, chr1 = "chr2L", chr2 = "chr2L",
resolution = 100000)

Container_file <- system.file(file.path("extdata"))
Brick_test <- load_BrickContainer(project_dir = Container_file)
Brick_matrix_dimensions(Brick = Brick_test, chr1 = "chr2L", chr2 = "chr2L",
resolution = 100000)

}
