% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_mcool_normalisation_exists}
\alias{Brick_mcool_normalisation_exists}
\title{Check if a normalisation exists in an mcool file.}
\usage{
Brick_mcool_normalisation_exists(mcool, norm.factor = NULL,
    binsize = NULL)
}
\arguments{
\item{mcool}{\strong{Required}.
Path to an mcool file.}

\item{norm.factor}{\strong{Optional}. Default "Iterative-Correction".
The normalization factor to use for normalization from an mcool file.
norm.factor currently accepts one of "Iterative-Correction", "Knight-Ruitz",
"Vanilla-coverage", "Vanilla-coverage-square-root"}

\item{binsize}{\strong{Optional}.
The binsize to select from an mcool file.}
}
\value{
A boolean vector of length 1
}
\description{
\code{Brick_mcool_normalisation_exists} checks if a particular normalisation
exists in an mcool file.
}
\examples{

\dontrun{

require(curl)
curl_download(url = paste("https://data.4dnucleome.org/"
"files-processed/4DNFI7JNCNFB/"
"@download/4DNFI7JNCNFB.mcool",sep = ""),
destfile = "./H1-hESC-HiC-4DNFI7JNCNFB.mcool")

mcool <- "./H1-hESC-HiC-4DNFI7JNCNFB.mcool"
Brick_mcool_normalisation_exists(mcool = mcool,
norm.factor = "Iterative-Correction",
binsize = 10000)

}

}
