% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_path_to_file}
\alias{Brick_path_to_file}
\title{Check if a Brick is being tracked.}
\usage{
Brick_path_to_file(Brick = NULL)
}
\arguments{
\item{Brick}{\strong{required}
Path to the Brick file to locate.}
}
\value{
Returns the corresponding path to the file that is being tracked.
}
\description{
\code{Brick_path_to_file} will query the tracking cache and return the complete
path to the file that is being tracked.
}
\examples{

Bintable.path <- system.file("extdata",
"Bintable_40kb.txt", package = "HiCBricks")
Chromosomes <- "chr19"
Path_to_cached_file <- CreateBrick(ChromNames = Chromosomes,
BinTable = Bintable.path, bin.delim = " ",
Output.Filename = "test.hdf", exec = "cat",
remove.existing = TRUE)

Brick.file <- "test.hdf"
Brick_path_to_file(Brick = Brick.file)
}
