% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_untrack_brick}
\alias{Brick_untrack_brick}
\title{Check if a Brick is being tracked.}
\usage{
Brick_untrack_brick(Brick = NULL)
}
\arguments{
\item{Brick}{\strong{required}
Path to the Brick file to untrack.}
}
\value{
Returns any value provided by bfcremove.
}
\description{
\code{Brick_untrack_brick} will stop tracking the provided Brick file,
if it exists.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
if(Brick_is_tracked(Brick.file)){
    Brick_untrack_brick(Brick = Brick.file)
}else{
    Brick_track_bricks(Brick = Brick.file)
    Brick_untrack_brick(Brick = Brick.file)
}
}
