% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{BrickContainer_change_experiment_name}
\alias{BrickContainer_change_experiment_name}
\title{Change the location of HDF files in the BrickContainer object}
\usage{
BrickContainer_change_experiment_name(Brick = NULL,
  experiment_name = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the BrickContainer created using
\code{Create_many_Bricks} or \code{Load_BrickContainer}}

\item{experiment_name}{\strong{Required}. Default NULL
A string specifying the new experiment name}
}
\value{
An object of class BrickContainer where the experiment_name
has been changed
}
\description{
\code{BrickContainer_change_experiment_name} changes the location of
name of the experiment
}
\examples{
Brick.file = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
BrickContainer_change_experiment_name(Brick = My_brick_container, 
experiment_name = "I change my mind")
}
