% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{BrickContainer_change_output_directory}
\alias{BrickContainer_change_output_directory}
\title{Change the location of HDF files in the BrickContainer object}
\usage{
BrickContainer_change_output_directory(Brick = NULL,
  output_directory = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{output_directory}{\strong{Required}. Default NULL
A string specifying new location of the output_directory. Please note,
that the location of the HDF files will not be changed.}
}
\value{
An object of class BrickContainer where the output_directory
has been changed
}
\description{
\code{BrickContainer_change_output_directory} changes the location of
associated HDF files
}
\examples{
Brick.file = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
BrickContainer_change_output_directory(Brick = My_brick_container, 
output_directory = tempdir())
}
