% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{BrickContainer_get_path_to_file}
\alias{BrickContainer_get_path_to_file}
\title{Get the path to HDF files present in the Brick container.}
\usage{
BrickContainer_get_path_to_file(Brick = NULL, chr1 = NA, chr2 = NA,
  type = NA, resolution = NA)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created
with Create_many_Bricks.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{type}{A value from one of cis, trans specifying the type of files to list
cis will list intra-choromosomal file paths and trans will list
inter-chromosomal file paths.}

\item{resolution}{\strong{Optional}. Default NA
When an object of class BrickContainer is provided, resolution defines the
resolution on which the function is executed}
}
\value{
A vector containing filepaths
}
\description{
\code{BrickContainer_get_path_to_file} fetches the list of HDF file paths
associated to a particular BrickContainer
}
\examples{
Brick.json = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
BrickContainer_get_path_to_file(Brick = My_brick_container)
}
