% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{BrickContainer_list_chromosomes}
\alias{BrickContainer_list_chromosomes}
\title{Return the descriptive name of the BrickContainer}
\usage{
BrickContainer_list_chromosomes(Brick = NULL, lengths = FALSE)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created
with Create_many_Bricks.}

\item{lengths}{Default FALSE
If TRUE, will also return the chromosomal lengths}
}
\value{
If lengths is FALSE, only the chromosome names are returned. If
lengths is TRUE, a data.frame containing the chromosome names and their
lengths is provided.
}
\description{
\code{BrickContainer_list_chromosomes} returns the chromosomes available in
the BrickContainer
}
\examples{
Brick.file = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
BrickContainer_list_chromosomes(My_brick_container)
}
