% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{BrickContainer_list_resolutions}
\alias{BrickContainer_list_resolutions}
\title{Return the descriptive name of the BrickContainer}
\usage{
BrickContainer_list_resolutions(Brick = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created
with Create_many_Bricks.}
}
\value{
A character string specifying the descriptive name of the
BrickContainer
}
\description{
\code{BrickContainer_list_resolutions} returns the resolutions available in
the BrickContainer
}
\examples{
Brick.file = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
BrickContainer_list_resolutions(My_brick_container)
}
