% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_add_ranges}
\alias{Lego_add_ranges}
\title{Perpetually store a ranges object in the Lego store.}
\usage{
Lego_add_ranges(Lego = NULL, ranges = NULL, rangekey = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{ranges}{\strong{Required}.
An object of class ranges specifying the ranges to store in the Lego.}

\item{rangekey}{\strong{Required}.
The name to use for the object within the Lego store.}
}
\description{
\code{Lego_add_ranges} loads a GRanges object into the Lego store.
}
\details{
With this function it is possible to associate other ranges objects with the
Lego store. If metadata columns are present, the are also loaded into the
Lego store. Although not explicitly asked for, the metadata columns should
not be of type list as this may create complications down the line. We
ask for ranges objects, so if the same ranges object is later retrieved
two additional columns will be present. These are the strand and width
columns that are obtained when a ranges is converted into a data.frame.
Users can ignore these columns.
}
