% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_fetch_range_index}
\alias{Lego_fetch_range_index}
\title{Returns the position of the supplied ranges in the binning table associated
to the Hi-C experiment.}
\usage{
Lego_fetch_range_index(Lego = NULL, chr = NULL, start = NULL,
  end = NULL, names = NULL, type = "any")
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr}{\strong{Required}.
A character vector of length N specifying the chromosomes to select from
the ranges.}

\item{start}{\strong{Required}.
A numeric vector of length N specifying the start positions in the chromosome}

\item{end}{\strong{Required}.
A numeric vector of length N specifying the end positions in the chromosome}

\item{names}{\strong{Optional}.
A character vector of length N specifying the names of the chromosomes. If
absent, names will take the form chr:start:end.}

\item{type}{\strong{Optional}. Default any
Type of overlap operation to do. It should be one of two, any or within.
any considers any overlap (atleast 1 bp) between the provided ranges and the
binning table.}
}
\value{
Returns a named list corresponding to one entry for each unique
chromosome present in the chr character vector. The list contains another
list with length equal to the number of ranges present per chromosome.
This list contains two named vectors, SubjectInfo and Indexes.
SubjectInfo contains a ranges created out of the vector from any given
position. Indexes contains a numeric vector corresponding to the rows/cols
overlapping the given ranges.
}
\description{
\code{Lego_fetch_range_index} constructs a ranges object and creates an overlap
operation between the constructed ranges and the Hi-C experiment associated
binning table finally returning a list of ranges with their corresponding
indices in the binning table.
}
