% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_fetch_row_vector}
\alias{Lego_fetch_row_vector}
\title{Return a matrix subset.}
\usage{
Lego_fetch_row_vector(Lego = NULL, chr1 = NULL, chr2 = NULL,
  by = NULL, vector = NULL, regions = NULL, flip = FALSE,
  FUN = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A string specifying the chromosome to subset on the rows}

\item{chr2}{\strong{Required}.
A string specifying the chromosome to subset on the columns}

\item{by}{\strong{Required}. One of two possible values, "position" or "ranges"
A one-dimensional numeric vector of length 1 specifying one of either
position or ranges.}

\item{vector}{\strong{Required}.
If by is position, a 1 dimensional numeric vector containing the rows to be
extracted is expected. If by is ranges, a 1 dimensional numeric vector
containing the names of the bintable is expected.
This function does not do overlaps. Rather it returns any given row or column
based on their position or names in the bintable.}

\item{regions}{\strong{Optional}.
A character vector of length vector is expected. Each element must be of the
form chr:start:end. These regions will be converted back to their original
positions and the corresponding rows will be subsetted by the corresponding
region element. If the length of regions does not match, the subset operation
will not be done and all elements from the rows will be returned.}

\item{flip}{\strong{Optional}. Default FALSE
If present, will flip everything. This is equivalent to selecting columns,
and subsetting on the rows.}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before the matrix
is returned.}
}
\value{
Returns a matrix of dimension x.vector length by y.vector length.
This may differ based on the operations with FUN.
}
\description{
\code{Lego_fetch_row_vector} will fetch any given rows from a matrix. If required,
the rows can be subsetted on the columns and transformations applied.
}
