% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_bintable}
\alias{Lego_get_bintable}
\title{Returns the binning table associated to the Hi-C experiment.}
\usage{
Lego_get_bintable(Lego = NULL, chr = NULL)
}
\description{
\code{Lego_get_bintable} makes a call to Lego_get_ranges to retrieve the
binning table of the associated Lego store. This is equivalent to passing
the argument rangekey = "bintable" in Lego_get_ranges
}
\seealso{
Lego_get_ranges
}
