% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_matrix}
\alias{Lego_get_matrix}
\title{Return a matrix subset.}
\usage{
Lego_get_matrix(Lego = NULL, chr1 = NULL, chr2 = NULL,
  x.vector = NULL, y.vector = NULL, FUN = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A string specifying the chromosome to subset on the rows}

\item{chr2}{\strong{Required}.
A string specifying the chromosome to subset on the columns}

\item{x.vector}{\strong{Required}.
A one-dimensional numeric vector specifying the the rows to subset.}

\item{y.vector}{\strong{Required}.
A one-dimensional numeric vector specifying the the columns to subset.}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before the matrix
is returned.}
}
\value{
Returns a matrix of dimension x.vector length by y.vector length.
This may differ based on the operations with FUN.
}
\description{
\code{Lego_get_matrix} will fetch a matrix subset between row values ranging from
min(x.vector) to max(x.vector) and column values ranging from min(x.vector)
to max(x.vector)
}
