% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_matrix_mcols}
\alias{Lego_get_matrix_mcols}
\title{Get the matrix metadata columns in the Lego store.}
\usage{
Lego_get_matrix_mcols(Lego = NULL, chr1 = NULL, chr2 = NULL,
    what = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{what}{\strong{Required}
A character vector of length 1 specifying the matrix metric to retrieve}
}
\value{
Returns a 1xN dimensional vector containing the specified matrix
metric
}
\description{
\code{Lego_get_matrix_mcols} will get the specified matrix metadata column.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_get_matrix_mcols(Lego = Lego.file, chr1 = "chr19", chr2 = "chr19",
what = "bin.coverage")
}
