% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_matrix_within_coords}
\alias{Lego_get_matrix_within_coords}
\title{Return a matrix subset between two regions.}
\usage{
Lego_get_matrix_within_coords(Lego = NULL, x.coords = NULL,
  y.coords = NULL, FUN = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{x.coords}{\strong{Required}.
A string specifying the region to subset on the rows. It takes the form
chr:start:end. An overlap operation with the associated bintable will be done
to identify the bins to subset on the row}

\item{y.coords}{\strong{Required}.
A string specifying the region to subset on the rows. It takes the form
chr:start:end. An overlap operation with the associated bintable will be done
to identify the bins to subset on the column}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before the matrix
is returned.}
}
\value{
Returns a matrix of dimension x.coords binned length by y.coords
binned length. This may differ based on FUN.
}
\description{
\code{Lego_get_matrix_within_coords} will fetch a matrix subset after creating an
overlap operation between both regions and the bintable associated to the
Lego store. This function calls Lego_get_matrix.
}
\seealso{
Lego_get_matrix
}
