% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_values_by_distance}
\alias{Lego_get_values_by_distance}
\title{Return values separated by a certain distance.}
\usage{
Lego_get_values_by_distance(Lego = NULL, chr = NULL, distance = NULL,
  constrain.region = NULL, batch.size = 500, FUN = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr}{\strong{Required}.
A string specifying the chromosome for the cis Hi-C matrix from which values
will be retrieved at a certain distance.}

\item{distance}{\strong{Required}. 0 based.
Fetch values separated by distance.}

\item{constrain.region}{\strong{Optional}.
A character vector of length 1 with the form chr:start:end specifying the
region for which the distance values must be retrieved.}

\item{batch.size}{\strong{Optional}. Default 500
A numeric vector of length 1 specifying the size of the chunk to retrieve
for diagonal selection.}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before values
are returned.}
}
\value{
Returns a numeric vector of length N depending on the presence of
constrain.region, FUN and distance from the main diagonal.
}
\description{
\code{Lego_get_values_by_distance} can fetch values with or without transformation
or subsetted by a certain distance.
}
