% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_is_tracked}
\alias{Lego_is_tracked}
\title{Check if a Lego is being tracked.}
\usage{
Lego_is_tracked(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{required}
Path to the Lego file to check.}
}
\value{
Returns TRUE or FALSE indicating if the file path is being tracked
or not.
}
\description{
\code{Lego_is_tracked} checks if the provided Lego file is being tracked or not.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos") 
Lego_is_tracked(Lego = Lego.file)
}
