% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_list_matrix_mcols}
\alias{Lego_list_matrix_mcols}
\title{List the matrix metadata columns in the Lego store.}
\usage{
Lego_list_matrix_mcols(Lego = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
columns of the matrix}
}
\value{
Returns a vector containing the names of all matrix metadata columns
}
\description{
\code{Lego_get_matrix_mcols} will list the names of all matrix metadata columns.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_list_matrix_mcols(Lego = Lego.file, chr1 = "chr19", chr2 = "chr19")
}
