% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_list_rangekeys}
\alias{Lego_list_rangekeys}
\title{List the ranges tables stored within the Lego.}
\usage{
Lego_list_rangekeys(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}
}
\value{
A one dimensional character vector of length x specifying the names
of all ranges currently present in the file.
}
\description{
\code{Lego_list_rangekeys} lists the names of all ranges associated to a Lego.
}
\examples{

Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_list_rangekeys(Lego = Lego.file)
 
}
