% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_list_ranges_mcols}
\alias{Lego_list_ranges_mcols}
\title{Find out what metadata columns are associated to a ranges with a certain name}
\usage{
Lego_list_ranges_mcols(Lego = NULL, rangekey = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{rangekey}{\strong{Optional}.
A string specifying the name of the ranges. If not present, the metadata
columns of all ranges will be listed.}
}
\value{
if no metadata columns are present, NA. If metadata columns are
present, a data.frame object containing the name of the ranges and the
associated metadata column name.
}
\description{
\code{Lego_list_ranges_mcols} will list the metadata columns of the specified
ranges if it is present in the Lego store.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_list_ranges_mcols(Lego = Lego.file, rangekey = "Bintable")

}
