% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_list_tracked_legos}
\alias{Lego_list_tracked_legos}
\title{List all Legos that have been created with .}
\usage{
Lego_list_tracked_legos(detailed = FALSE, preserve.col.names = FALSE)
}
\arguments{
\item{detailed}{\strong{Optional}
If FALSE, produces a data.frame of file paths pointing to the
Lego objects. If TRUE, produces a tibble containing additional information
such as the bioc cache tracking id, and the creation, last accession and
modification times.}

\item{preserve.col.names}{\strong{Optional}
If TRUE, will preserve the original colnames from the bioc cache tibble
object. If FALSE, will attempt to humanize the colnames to improve
readability.}
}
\value{
Returns a data.frame or tibble containing the path to tracked
Lego objects. If tibble will contain additional information related to
the tracking. For details see parameter detailed.
}
\description{
\code{Lego_list_tracked_legos} will list all Lego objects that are 
being tracked.
}
\examples{
Lego_list_tracked_legos()
}
