% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_load_matrix}
\alias{Lego_load_matrix}
\title{Load a NxM dimensional matrix into the Lego store.}
\usage{
Lego_load_matrix(Lego = NULL, chr1 = NULL, chr2 = NULL,
  matrix.file = NULL, delim = " ", exec = NULL,
  remove.prior = FALSE, num.rows = 2000, is.sparse = FALSE,
  distance = NULL, sparsity.bins = 100)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
columns of the matrix}

\item{matrix.file}{\strong{Required}.
A character vector of length 1 specifying the name of the file to load as a
matrix into the Lego store.}

\item{delim}{\strong{Optional}. Default " "
The delimiter of the matrix file.}

\item{exec}{\strong{Required}.
A string specifying the program to use for reading the file. Use cat for txt
files, for bz2 files use bzcat and for gz files zcat.}

\item{num.rows}{\strong{Optional}. Default 2000
Number of rows to read, in each chunk.}

\item{is.sparse}{\strong{Optional}. Default FALSE
If true, designates the matrix as being a sparse matrix, and computes the
sparsity.index. The sparsity index measures the proportion of non-zero rows
or columns at a certain distance from the diagonal (100) in cis interaction
matrices.}

\item{distance}{\strong{Optional}. Default NULL. Not implemented yet.
For very high-resolution matrices, read times can become extremely slow and
it does not make sense to load the entire matrix into the data structure, as
after a certain distance, the matrix will become extremely sparse. If
provided, only interactions upto a certain distance from the main diagonal
will be loaded into the data structure.}

\item{sparsity.bins}{\strong{Optional}. Default 100
With regards to computing the sparsity.index, this parameter decides the
number of bins to scan from the diagonal.}

\item{remove.priori}{\strong{Optional}. Default FALSE
If a matrix was loaded before, it will not be replaced. Use remove.priori to
override and replace the existing matrix.}
}
\description{
Load a NxM dimensional matrix into the Lego store.
}
