% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_matrix_filename}
\alias{Lego_matrix_filename}
\title{Return the filename of the loaded matrix}
\usage{
Lego_matrix_filename(Lego = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to the
columns of the matrix}
}
\value{
Returns a character vector of length 1 specifying the initial
filename of the matrix.
}
\description{
Return the filename of the loaded matrix
}
