% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_path_to_file}
\alias{Lego_path_to_file}
\title{Check if a Lego is being tracked.}
\usage{
Lego_path_to_file(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{required}
Path to the Lego file to locate.}
}
\value{
Returns the corresponding path to the file that is being tracked.
}
\description{
\code{Lego_path_to_file} will query the tracking cache and return the complete
path to the file that is being tracked.
}
\examples{
Lego.file <- "test.hdf"
Lego_path_to_file(Lego = Lego.file)
}
