% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_rangekey_exists}
\alias{Lego_rangekey_exists}
\title{Check to see if the Lego contains a ranges with a certain name.}
\usage{
Lego_rangekey_exists(Lego = NULL, rangekey = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{rangekey}{\strong{Required}.
A string specifying the name of the ranges to check for.}
}
\value{
A logical vector of length 1 with either TRUE or FALSE values.
}
\description{
\code{Lego_rangekey_exists} checks for the presence of a particular ranges with
a certain name.
}
\examples{

Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_rangekey_exists(Lego = Lego.file, rangekey = "Bintable")

}
