% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_return_region_position}
\alias{Lego_return_region_position}
\title{Provides the corresponding overlapping position from the bintable.}
\usage{
Lego_return_region_position(Lego = NULL, region = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{region}{\strong{Required}.
A character vector of length 1 specifying the region to overlap. It must take
the form chr:start:end.}
}
\value{
Returns a 1 dimensional vector containing the position of the
overlapping regions in the bintable associated the Lego store.
}
\description{
\code{Lego_return_region_position} takes as input a human-readable coordinate
format of the form chr:start:end and outputs the overlapping bintable
positions.
}
