% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_track_legos}
\alias{Lego_track_legos}
\title{Check if a Lego is being tracked.}
\usage{
Lego_track_legos(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{required}
Path to the Lego file to track.}
}
\value{
Returns a named vector containing the path to the file along with
the bioc Cache id as its name.
}
\description{
\code{Lego_track_legos} will start tracking the provided Lego file.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos") 
Lego_track_legos(Lego = Lego.file)
}
