% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_untrack_lego}
\alias{Lego_untrack_lego}
\title{Check if a Lego is being tracked.}
\usage{
Lego_untrack_lego(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{required}
Path to the Lego file to untrack.}
}
\description{
\code{Lego_untrack_lego} will stop tracking the provided Lego file, if it exists.
}
\examples{
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos") 
Lego_untrack_lego(Lego = Lego.file)
}
