% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrickContainer_functions.R
\name{load_BrickContainer}
\alias{load_BrickContainer}
\title{Create a BrickContainer object from a JSON file}
\usage{
load_BrickContainer(config_file = NULL, project_dir = NULL)
}
\arguments{
\item{config_file}{Default NULL
A character string of length 1 specifying the path to the path to the
configuration json created using \code{Create_many_bricks}}

\item{project_dir}{Default NULL
A character string of length 1 specifying the path to the path to the
configuration json created using \code{Create_many_bricks}}
}
\value{
An object of class BrickContainer
}
\description{
\code{load_BrickContainer} creates a BrickContainer object from a JSON file
}
\examples{
Brick.file = system.file("extdata", "test.json", package = "HiCBricks")
My_brick_container <- load_BrickContainer(Brick.file)
}
