% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_bintable}
\alias{Brick_get_bintable}
\title{Returns the binning table associated to the Hi-C experiment.}
\usage{
Brick_get_bintable(Brick = NULL, chr = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr}{\strong{Optional}.
A chr string specifying the chromosome to select from the ranges.}
}
\value{
Returns a GRanges object containing the binning
table associated to the Brick store.
}
\description{
\code{Brick_get_bintable} makes a call to \code{\link{Brick_get_ranges}} to
retrieve the binning table of the associated Brick store. This is equivalent
to passing the argument rangekey = "bintable" in
\code{\link{Brick_get_ranges}}
}
\examples{

Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_get_bintable(Brick = Brick.file)

}
\seealso{
Brick_get_ranges
}
