% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_chrominfo}
\alias{Brick_get_chrominfo}
\title{Get the chrominfo for the Hi-C experiment.}
\usage{
Brick_get_chrominfo(Brick = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}
}
\value{
A three column data.frame containing chromosomes, nrows and length.

chromosomes corresponds to all chromosomes in the provided bintable.

nrows corresponds to the number of entries in the bintable or dimension
for that chromosome in a Hi-C matrix.

Length is the total bp length of the same
chromosome (max value for that chromosome in the bintable).
}
\description{
\code{Brick_get_chrominfo} fetches the associated chrominfo table for the
Brick it is associated to.
}
\examples{
Brick.file = system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_get_chrominfo(Brick = Brick.file)
}
