% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_matrix_mcols}
\alias{Brick_get_matrix_mcols}
\title{Get the matrix metadata columns in the Brick store.}
\usage{
Brick_get_matrix_mcols(Brick, chr1, chr2, resolution, what)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with
Create_many_Brick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{resolution}{\strong{Optional}. Default NA
When an object of class BrickContainer is provided, resolution defines the
resolution on which the function is executed}

\item{what}{\strong{Required}
A character vector of length 1 specifying the matrix metric to retrieve}
}
\value{
Returns a 1xN dimensional vector containing the specified matrix
metric
}
\description{
\code{Brick_get_matrix_mcols} will get the specified matrix metadata column.
}
\examples{
Container_file <- system.file(file.path("extdata",
"HiCBricks_builder_config.json"), package = "HiCBricks")
Brick_test <- load_BrickContainer(Container_file)
Brick_get_matrix_mcols(Brick = Brick_test, chr1 = "chr2L", chr2 = "chr2L",
resolution = 100000, what = "bin.coverage")
}
