% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_matrix_mcols}
\alias{Brick_get_matrix_mcols}
\title{Get the matrix metadata columns in the Brick store.}
\usage{
Brick_get_matrix_mcols(Brick = NULL, chr1 = NULL, chr2 = NULL,
    what = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{what}{\strong{Required}
A character vector of length 1 specifying the matrix metric to retrieve}
}
\value{
Returns a 1xN dimensional vector containing the specified matrix
metric
}
\description{
\code{Brick_get_matrix_mcols} will get the specified matrix metadata column.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_get_matrix_mcols(Brick = Brick.file, chr1 = "chr19", chr2 = "chr19",
what = "bin.coverage")
}
