% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_get_values_by_distance}
\alias{Brick_get_values_by_distance}
\title{Return values separated by a certain distance.}
\usage{
Brick_get_values_by_distance(Brick, chr, distance, constrain.region = NULL,
    batch.size = 500, FUN = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr}{\strong{Required}.
A string specifying the chromosome for the cis Hi-C matrix from which values
will be retrieved at a certain distance.}

\item{distance}{\strong{Required}. 0 based.
Fetch values separated by distance.}

\item{constrain.region}{\strong{Required}.
A character vector of length 1 with the form chr:start:end specifying the
region for which the distance values must be retrieved.}

\item{batch.size}{\strong{Optional}. Default 500
A numeric vector of length 1 specifying the size of the chunk to retrieve
for diagonal selection.}

\item{FUN}{\strong{Optional}.
If provided a data transformation with FUN will be applied before values
are returned.}
}
\value{
Returns a numeric vector of length N depending on the presence of
constrain.region, FUN and distance from the main diagonal.
}
\description{
\code{Brick_get_values_by_distance} can fetch values with or without
transformation or subsetted by a certain distance. Please note,
this module is not an iterable module.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_get_values_by_distance(Brick = Brick.file, chr = "chr19",
distance = 0)

Failsafe_median <- function(x){
    x[is.nan(x) | is.infinite(x) | is.na(x)] <- 0
    return(median(x))
}

Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_get_values_by_distance(Brick = Brick.file, chr = "chr19",
distance = 4, FUN = Failsafe_median)

}
\seealso{
\code{\link{Brick_get_matrix_within_coords}} to get matrix by
using matrix coordinates, \code{\link{Brick_fetch_row_vector}} to get values
in a certain row/col and subset them, \code{\link{Brick_get_vector_values}}
to get values using matrix coordinates, \code{\link{Brick_get_matrix}} to
get matrix by using matrix coordinates.
}
