% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_list_matrix_mcols}
\alias{Brick_list_matrix_mcols}
\title{List the matrix metadata columns in the Brick store.}
\usage{
Brick_list_matrix_mcols()
}
\value{
Returns a vector containing the names of all matrix metadata columns
}
\description{
\code{Brick_get_matrix_mcols} will list the names of all matrix metadata
columns.
}
\examples{
Container_file <- system.file(file.path("extdata",
"HiCBricks_builder_config.json"), package = "HiCBricks")
Brick_test <- load_BrickContainer(Container_file)
Brick_list_matrix_mcols()
}
