% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_list_mcool_resolutions}
\alias{Brick_list_mcool_resolutions}
\title{Get all available normalisations in an mcool file.}
\usage{
Brick_list_mcool_resolutions(mcool = NULL)
}
\arguments{
\item{mcool}{\strong{Required}.
A parameter specifying the name of an mcool file}
}
\value{
A named vector listing all possible resolutions in the file.
}
\description{
\code{Brick_list_mcool_resolutions} lists all available resolutions in the mcool
file.
}
