% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_list_rangekeys}
\alias{Brick_list_rangekeys}
\title{List the ranges tables stored within the Brick.}
\usage{
Brick_list_rangekeys(Brick)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}
}
\value{
A one dimensional character vector of length x specifying the names
of all ranges currently present in the file.
}
\description{
\code{Brick_list_rangekeys} lists the names of all ranges associated to a Brick.
}
\examples{

Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_list_rangekeys(Brick = Brick.file)

}
