% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_matrix_dimensions}
\alias{Brick_matrix_dimensions}
\title{Return the dimensions of a matrix}
\usage{
Brick_matrix_dimensions(Brick = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns the dimensions of a Hi-C matrix for any given
chromosome pair.
}
\description{
Return the dimensions of a matrix
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_matrix_dimensions(Brick = Brick.file, chr1 = "chr19", chr2 = "chr19")

}
