% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_matrix_exists}
\alias{Brick_matrix_exists}
\title{Check if a chromosome pair exists.}
\usage{
Brick_matrix_exists(Brick, chr1, chr2, resolution = NA)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with
Create_many_Brick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}

\item{resolution}{\strong{Optional}. Default NA
When an object of class BrickContainer is provided, resolution defines the
resolution on which the function is executed}
}
\value{
Returns a logical vector of length 1, specifying if the matrix
exists or not.
}
\description{
Matrices are created when the bintable is loaded and the chromosome names
are provided. If a user is in doubt regarding whether a matrix is present or
not it is useful to check this function. If the Bintable did not contain a
particular chromosome, any matrices for that chromosome would not be present
in the file
}
\examples{
Container_file <- system.file(file.path("extdata",
"HiCBricks_builder_config.json"), package = "HiCBricks")
Brick_test <- load_BrickContainer(Container_file)
Brick_matrix_exists(Brick = Brick_test, chr1 = "chr2L", chr2 = "chr2L",
resolution = 100000)

Container_file <- system.file(file.path("extdata"))
Brick_test <- load_BrickContainer(project_dir = Container_file)
Brick_matrix_exists(Brick = Brick_test, chr1 = "chr2L", chr2 = "chr2L",
resolution = 100000)

}
