% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_matrix_issparse}
\alias{Brick_matrix_issparse}
\title{Check if a matrix for a chromosome pair is sparse.}
\usage{
Brick_matrix_issparse(Brick = NULL, chr1 = NULL, chr2 = NULL)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns a logical vector of length 1, specifying if a matrix was
loaded as a sparse matrix.
}
\description{
Check if a matrix for a chromosome pair is sparse.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_matrix_issparse(Brick = Brick.file, chr1 = "chr19", chr2 = "chr19")

}
