% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_matrix_maxdist}
\alias{Brick_matrix_maxdist}
\title{Get the maximum loaded distance from the diagonal of any matrix.}
\usage{
Brick_matrix_maxdist(Brick, chr1, chr2)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{chr1}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the rows of the matrix}

\item{chr2}{\strong{Required}.
A character vector of length 1 specifying the chromosome corresponding to
the columns of the matrix}
}
\value{
Returns an integer vector of length 1, specifying the maximum
distance loaded for that matrix
}
\description{
If values beyond a certain distance were not loaded in the matrix, this
distance parameter is useful. This package by default will check this param
to make sure that it is not returning non-existent data.
}
\details{
\code{Brick_matrix_maxdist} will return this parameter.
}
\examples{
Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_matrix_maxdist(Brick = Brick.file, chr1 = "chr19", chr2 = "chr19")

}
