% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brick_functions.R
\name{Brick_rangekey_exists}
\alias{Brick_rangekey_exists}
\title{Check to see if the Brick contains a ranges with a certain name.}
\usage{
Brick_rangekey_exists(Brick, rangekey)
}
\arguments{
\item{Brick}{\strong{Required}.
A string specifying the path to the Brick store created with CreateBrick.}

\item{rangekey}{\strong{Required}.
A string specifying the name of the ranges to check for.}
}
\value{
A logical vector of length 1 with either TRUE or FALSE values.
}
\description{
\code{Brick_rangekey_exists} checks for the presence of a particular ranges with
a certain name.
}
\examples{

Brick.file <- system.file("extdata", "test.hdf", package = "HiCBricks")
Brick_rangekey_exists(Brick = Brick.file, rangekey = "Bintable")

}
