% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_fetch_range_index}
\alias{Lego_fetch_range_index}
\title{Returns the position of the supplied ranges in the binning table
associated to the Hi-C experiment.}
\usage{
Lego_fetch_range_index(Lego = NULL, chr = NULL, start = NULL,
    end = NULL, names = NULL, type = "any")
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr}{\strong{Required}.
A character vector of length N specifying the chromosomes to select from
the ranges.}

\item{start}{\strong{Required}.
A numeric vector of length N specifying the start positions in the
chromosome}

\item{end}{\strong{Required}.
A numeric vector of length N specifying the end positions in the chromosome}

\item{names}{\strong{Optional}.
A character vector of length N specifying the names of the chromosomes. If
absent, names will take the form chr:start:end.}

\item{type}{\strong{Optional}. Default any
Type of overlap operation to do. It should be one of two, any or within.
any considers any overlap (atleast 1 bp) between the provided ranges and
the binning table.}
}
\value{
Returns a GenomicRanges object of same length as the chr, start, end
vectors provided. The object is returned with an additional column, Indexes.
Indexes is a column of class \code{\link[IRanges]{IntegerList}}, which is
part of the larger \code{\link[IRanges]{AtomicList}} superset. This
"Indexes" column can be accessed like a normal GRanges column with the
additional list accessor [[]] in place of the normal vector accessor [].
}
\description{
\code{Lego_fetch_range_index} constructs a ranges object using
\code{\link{Lego_make_ranges}}, creates an overlap operation using
\code{\link[GenomicRanges]{findOverlaps}}, where the constructed ranges is
the \emph{subject} and the Hi-C experiment associated binning table is the
\emph{query}. The return of this object is a list of ranges with their
corresponding indices in the binning table.
}
\examples{

Chrom <- c("chr19","chr19")
Start <- c(1,40000)
End <- c(1000000,2000000)
Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Test_Run <- Lego_fetch_range_index(Lego = Lego.file, chr = Chrom, 
start = Start, end = End)
Test_Run$Indexes[[1]]

}
