% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_bintable}
\alias{Lego_get_bintable}
\title{Returns the binning table associated to the Hi-C experiment.}
\usage{
Lego_get_bintable(Lego = NULL, chr = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}

\item{chr}{\strong{Optional}.
A chr string specifying the chromosome to select from the ranges.}
}
\value{
Returns a GRanges object containing the binning table associated to the
Lego store.
}
\description{
\code{Lego_get_bintable} makes a call to \code{\link{Lego_get_ranges}} to
retrieve the binning table of the associated Lego store. This is equivalent
to passing the argument rangekey = "bintable" in
\code{\link{Lego_get_ranges}}
}
\examples{

Lego.file <- system.file("extdata", "test.hdf", package = "HiCLegos")
Lego_get_bintable(Lego = Lego.file)

}
\seealso{
Lego_get_ranges
}
