% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lego_functions.R
\name{Lego_get_chrominfo}
\alias{Lego_get_chrominfo}
\title{Get the chrominfo for the Hi-C experiment.}
\usage{
Lego_get_chrominfo(Lego = NULL)
}
\arguments{
\item{Lego}{\strong{Required}.
A string specifying the path to the Lego store created with CreateLego.}
}
\value{
A three column file containing chromosomes, nrows and length.
chromosomes corresponds to all chromosomes in the provided bintable, nrows
corresponds to the number of entries in the bintable or dimension for that
chromosome in the Hi-C matrix. And length is the total bp length of the same
chromosome (max value for that chromosome in the bintable).
}
\description{
\code{Lego_get_chrominfo} fetches the associated chrominfo table for the
Lego (HDF) it is associated to.
}
